Разбивает проект на js модули
